adding more funcitonality to new dashboard
added dropdown to three dot more option component, connected header text to correct parts of the website, added permissions tabs to dashboard mine, shared, shared with me for documents and annotations. Connected New Group and New document button to correct link. Added box-shadow to footer so that it would be visible on the dashboard because I changed the color of the background color of the dashboard