added more filter functionality and made new component SplashPage to make index.js file less crowded
made a splashpage component to make the index.js file less crowded.
Made filters that are selected that have 0 results still clickable so the user can deselect the filter in the Filter dropdown then once it is deselected then it becomes disabled.
Added code to make sure all the document data we have gets populated into the IS Annotation Filter